fixing shortened helper, moved to a new package and integrated also in /create/check response into a json
