+++ election.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/vocdoni/vote-frame/mongo"
+	"github.com/vocdoni/vote-frame/shortener"
 	"go.vocdoni.io/dvote/api"
 	"go.vocdoni.io/dvote/apiclient"
 	"go.vocdoni.io/dvote/httprouter"
@@ -22,9 +23,6 @@ import (
 const (
 	ElectionSourceWebApp = "farcaster.vote"
 	ElectionSourceBot    = "bot"
-	shortenerTimeout     = 10 * time.Second
-	shortenerBase        = "https://frame.vote/"
-	shortenerEndpoint    = shortenerBase + "add/%s"
 )
 
 func (v *vocdoniHandler) election(electionID types.HexBytes) (*api.Election, error) {
@@ -136,7 +134,16 @@ func (v *vocdoniHandler) checkElection(msg *apirest.APIdata, ctx *httprouter.HTT
 	if !ok {
 		return ctx.Send(nil, http.StatusNoContent)
 	}
-	return ctx.Send(nil, http.StatusOK)
+	frameUrl := fmt.Sprintf("%s/%x", serverURL, electionID)
+	shortedURL, err := shortener.ShortURL(ctx.Request.Context(), frameUrl)
+	if err != nil {
+		return fmt.Errorf("failed to shorten election url: %w", err)
+	}
+	body, err := json.Marshal(map[string]string{"url": shortedURL})
+	if err != nil {
+		return fmt.Errorf("failed to marshal response: %w", err)
+	}
+	return ctx.Send(body, http.StatusOK)
 }
 
 // votersForElection returns the list of voters for the given election.
@@ -346,33 +353,4 @@ func (v *vocdoniHandler) saveElectionAndProfile(election *api.Election, profile 
 		return fmt.Errorf("failed to update user in database: %w", err)
 	}
 	return nil
-}
-
-// ShortElectionURL returns a shortened version of the election URL. It uses the
-// vocdoni shortener service to shorten the URL. It returns the shortened URL or
-// an error if something went wrong. It uses a timeout of 10 seconds for the
-// request.
-func ShortElectionURL(ctx context.Context, electionURL string) (string, error) {
-	internalCtx, cancel := context.WithTimeout(ctx, shortenerTimeout)
-	defer cancel()
-	endpointURL := fmt.Sprintf(shortenerEndpoint, electionURL)
-	req, err := http.NewRequestWithContext(internalCtx, http.MethodGet, endpointURL, nil)
-	if err != nil {
-		return "", fmt.Errorf("failed to create request: %w", err)
-	}
-	res, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return "", fmt.Errorf("failed to send request: %w", err)
-	}
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("error downloading json: %s", res.Status)
-	}
-	shortenerResponse := &struct {
-		Link string `json:"link"`
-	}{}
-	if err := json.NewDecoder(res.Body).Decode(shortenerResponse); err != nil {
-		return "", fmt.Errorf("failed to decode json: %w", err)
-	}
-	return shortenerBase + shortenerResponse.Link, nil
 }
